refactor: extract note components
