+++ src/components/note-sheet-content.tsx
@@ -1,684 +1,16 @@
-
 "use client";
 
-import { useEffect, useState, useRef } from 'react';
-import { Camera, Heart, Flag, Send, X, Trash2 } from 'lucide-react';
-import { Note, Reply } from '@/types';
-import { Button } from './ui/button';
-import { Textarea } from './ui/textarea';
-import { Badge } from './ui/badge';
-import { Avatar, AvatarFallback } from './ui/avatar';
-import { Separator } from './ui/separator';
-import { useToast } from '@/hooks/use-toast';
-import { ScrollArea } from './ui/scroll-area';
-import { Coordinates } from '@/hooks/use-location';
-import { doc, getDoc, Timestamp, collection, onSnapshot, query, orderBy, addDoc, serverTimestamp, runTransaction, where, deleteDoc, updateDoc, writeBatch } from 'firebase/firestore';
-import { db, storage } from '@/lib/firebase';
-import { getOrCreatePseudonym } from '@/lib/pseudonym';
-import { getDownloadURL, ref, uploadBytes } from 'firebase/storage';
-import { Skeleton } from './ui/skeleton';
-import { useAuth } from './auth-provider';
-import { z } from 'zod';
-import { moderateContent } from '@/ai/flows/content-moderation';
-import { Input } from './ui/input';
-import Image from 'next/image';
-import { cn } from '@/lib/utils';
-import { geohashForLocation } from 'geofire-common';
-import {
-  AlertDialog,
-  AlertDialogAction,
-  AlertDialogCancel,
-  AlertDialogContent,
-  AlertDialogDescription,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogTitle,
-} from "@/components/ui/alert-dialog";
-import { Label } from './ui/label';
-import { SheetFooter, SheetHeader, SheetTitle } from './ui/sheet';
-import { calculateReportUpdate } from '@/lib/reporting';
+import { useEffect, useState } from "react";
+import { doc, getDoc, Timestamp } from "firebase/firestore";
 
+import { useToast } from "@/hooks/use-toast";
+import { Coordinates } from "@/hooks/use-location";
+import { db } from "@/lib/firebase";
+import { Note } from "@/types";
 
-function SubmitButton({label, pendingLabel, isSubmitting}: {label: string, pendingLabel: string, isSubmitting: boolean}) {
-  return (
-    <Button type="submit" disabled={isSubmitting}>
-      {isSubmitting ? (
-        <>
-          <div className="animate-spin rounded-full h-4 w-4 border-b-2 border-current mr-2"></div>
-          {pendingLabel}
-        </>
-      ) : label}
-    </Button>
-  );
-}
-
-const noteSchema = z.object({
-  text: z.string().min(1, "Note cannot be empty.").max(800, "Note cannot exceed 800 characters."),
-  image: z.instanceof(File).nullish(),
-});
-
-
-function CreateNoteForm({ userLocation, onNoteCreated, onClose }: { userLocation: Coordinates | null, onNoteCreated: () => void, onClose: () => void }) {
-    const { user } = useAuth();
-    const { toast } = useToast();
-    const formRef = useRef<HTMLFormElement>(null);
-    const fileInputRef = useRef<HTMLInputElement>(null);
-    const [imagePreview, setImagePreview] = useState<string | null>(null);
-    const [imageFile, setImageFile] = useState<File | null>(null);
-    const [imageDimensions, setImageDimensions] = useState<{ w: number; h: number } | null>(null);
-    const [isSubmitting, setIsSubmitting] = useState(false);
-    
-    const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
-        e.preventDefault();
-        if (!user || !userLocation) {
-            toast({ title: "Error", description: "User or location not available.", variant: "destructive" });
-            return;
-        }
-
-        setIsSubmitting(true);
-        const formData = new FormData(e.currentTarget);
-        const text = formData.get('text') as string;
-
-        try {
-            // 1. Validate content
-            const validation = noteSchema.safeParse({ text, image: imageFile });
-            if (!validation.success) {
-                toast({ title: "Invalid Input", description: validation.error.errors[0].message, variant: "destructive" });
-                setIsSubmitting(false);
-                return;
-            }
-
-            // 2. Get user pseudonym
-            const pseudonym = await getOrCreatePseudonym(user.uid);
-            
-            // 3. Calculate geohash
-            const geohash = geohashForLocation([userLocation.latitude, userLocation.longitude]);
-
-            // 4. Prepare base note data
-             const newNote: Omit<Note, 'id' | 'createdAt'> & { createdAt: any } = {
-              text,
-              lat: userLocation.latitude,
-              lng: userLocation.longitude,
-              geohash,
-              authorUid: user.uid,
-              createdAt: serverTimestamp(),
-              authorPseudonym: pseudonym,
-              type: imageFile ? 'photo' : 'text',
-              score: 0,
-              teaser: text.substring(0, 30) + (text.length > 30 ? '...' : ''),
-              visibility: 'public',
-              trust: 0.5,
-              placeMaskMeters: 10,
-              revealMode: 'proximity+sightline',
-              revealRadiusM: 35,
-              revealAngleDeg: 20,
-              peekable: false,
-              dmAllowed: false,
-              media: [],
-            };
-
-            // 5. Handle image upload if present
-            if (imageFile) {
-                if (imageFile.size > 5 * 1024 * 1024) { // 5MB limit
-                    toast({ title: "Image too large", description: "Image must be less than 5MB.", variant: "destructive" });
-                    setIsSubmitting(false);
-                    return;
-                }
-                const storageRef = ref(storage, `notes/${user.uid}/${Date.now()}-${imageFile.name}`);
-                
-                const UPLOAD_TIMEOUT = 30000; // 30 seconds
-                const uploadPromise = uploadBytes(storageRef, imageFile);
-                
-                const timeoutPromise = new Promise((_, reject) => 
-                    setTimeout(() => reject(new Error("Upload timeout")), UPLOAD_TIMEOUT)
-                );
-
-                try {
-                    const snapshot = await Promise.race([uploadPromise, timeoutPromise]);
-                    const downloadURL = await getDownloadURL((snapshot as any).ref);
-
-                    newNote.media = [{
-                        path: downloadURL,
-                        type: 'image',
-                        w: imageDimensions?.w ?? 0,
-                        h: imageDimensions?.h ?? 0,
-                    }];
-                } catch (uploadError: any) {
-                    if (uploadError.message === "Upload timeout") {
-                         toast({ title: "Image upload timed out", description: "Please check your network connection and try again.", variant: "destructive" });
-                    } else {
-                        toast({ title: "Image upload failed", description: "Could not upload the image. Please try again.", variant: "destructive" });
-                    }
-                    console.error("Upload failed:", uploadError);
-                    setIsSubmitting(false);
-                    return;
-                }
-            }
-
-            // 6. Save note to Firestore
-            await addDoc(collection(db, 'notes'), newNote);
-            
-            toast({ title: "Success!", description: "Note dropped successfully!" });
-            formRef.current?.reset();
-            setImagePreview(null);
-            setImageFile(null);
-            onNoteCreated(); // This will fetch notes and close the sheet
-        } catch (error: any) {
-            console.error("Error creating note:", error);
-            toast({ title: "Error creating note", description: error.message || "An unknown error occurred.", variant: "destructive" });
-        } finally {
-            setIsSubmitting(false);
-        }
-    }
-    
-    const handleImageChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-        const file = event.target.files?.[0];
-        if (file) {
-            setImageFile(file);
-            const reader = new FileReader();
-            reader.onloadend = () => {
-                const result = reader.result as string;
-                setImagePreview(result);
-                const img = new window.Image();
-                img.onload = () => {
-                    setImageDimensions({ w: img.naturalWidth, h: img.naturalHeight });
-                };
-                img.src = result;
-            };
-            reader.readAsDataURL(file);
-        } else {
-            setImagePreview(null);
-            setImageFile(null);
-            setImageDimensions(null);
-        }
-    };
-    
-    const handleRemoveImage = () => {
-        setImagePreview(null);
-        setImageFile(null);
-        setImageDimensions(null);
-        if(fileInputRef.current) {
-            fileInputRef.current.value = '';
-        }
-    }
-
-    if (!user) {
-      return (
-        <div className="p-4 text-center text-muted-foreground">
-          You must be signed in to drop a note.
-        </div>
-      );
-    }
-
-    return (
-        <>
-            <SheetHeader>
-                <SheetTitle className="font-headline text-2xl">Drop a New Note</SheetTitle>
-            </SheetHeader>
-            <ScrollArea className="flex-1 -mx-6">
-                <form ref={formRef} onSubmit={handleSubmit} className="px-6 pt-4 space-y-4">
-                    <Textarea name="text" placeholder="What's on your mind? (Max 800 chars)" maxLength={800} rows={5} required />
-                    
-                    {imagePreview && (
-                        <div className="relative">
-                            <Image width={400} height={300} src={imagePreview} alt="Image preview" className="rounded-md w-full object-cover aspect-video" />
-                            <Button type="button" variant="destructive" size="icon" className="absolute top-2 right-2 h-6 w-6" onClick={handleRemoveImage}>
-                                <X className="h-4 w-4" />
-                            </Button>
-                        </div>
-                    )}
-                     {userLocation && <p className="text-xs text-muted-foreground">Location: {userLocation.latitude.toFixed(5)}, {userLocation.longitude.toFixed(5)}</p>}
-                </form>
-            </ScrollArea>
-             <SheetFooter className="pt-4">
-                 <div className="flex items-center justify-between w-full">
-                     <Button variant="outline" size="icon" type="button" onClick={() => fileInputRef.current?.click()}>
-                        <Camera className="h-4 w-4"/>
-                    </Button>
-                    <Input
-                        type="file"
-                        name="image"
-                        className="hidden"
-                        ref={fileInputRef}
-                        onChange={handleImageChange}
-                        accept="image/png, image/jpeg, image/gif"
-                    />
-                    <Button form="create-note-form" type="submit" disabled={isSubmitting}>
-                        {isSubmitting ? (
-                            <>
-                            <div className="animate-spin rounded-full h-4 w-4 border-b-2 border-current mr-2"></div>
-                            Dropping...
-                            </>
-                        ) : 'Drop Note'}
-                    </Button>
-                </div>
-            </SheetFooter>
-        </>
-    );
-}
-
-const replySchema = z.object({
-  text: z.string().min(1, "Reply cannot be empty.").max(120, "Reply cannot exceed 120 characters."),
-});
-
-function ReplyForm({ noteId, noteAuthorUid }: { noteId: string, noteAuthorUid?: string | null }) {
-     const { user } = useAuth();
-     const { toast } = useToast();
-     const [text, setText] = useState("");
-     const [isSubmitting, setIsSubmitting] = useState(false);
-     const [error, setError] = useState<string | null>(null);
-
-    const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
-        e.preventDefault();
-        if (!user) {
-            toast({ title: "Not signed in", description: "You must be signed in to reply.", variant: "destructive" });
-            return;
-        }
-
-        const validation = replySchema.safeParse({ text });
-        if (!validation.success) {
-            setError(validation.error.errors[0].message);
-            return;
-        }
-        
-        setError(null);
-        setIsSubmitting(true);
-
-        try {
-             const pseudonym = await getOrCreatePseudonym(user.uid);
-             const replyRef = collection(db, 'notes', noteId, 'replies');
-
-             await addDoc(replyRef, {
-                 text,
-                 noteId,
-                 authorUid: user.uid,
-                 authorPseudonym: pseudonym,
-                 createdAt: serverTimestamp(),
-             });
-
-             if (noteAuthorUid && noteAuthorUid !== user.uid) {
-                 const notificationsRef = collection(db, 'notifications');
-                 await addDoc(notificationsRef, {
-                     userId: noteAuthorUid,
-                     type: 'reply',
-                     noteId,
-                     actorUid: user.uid,
-                     actorPseudonym: pseudonym,
-                     createdAt: serverTimestamp(),
-                     read: false,
-                 });
-                 fetch('/api/notify', {
-                     method: 'POST',
-                     headers: { 'Content-Type': 'application/json' },
-                     body: JSON.stringify({
-                         userId: noteAuthorUid,
-                         title: 'New reply',
-                         body: `${pseudonym} replied to your note`,
-                         data: { noteId, type: 'reply' }
-                     })
-                 }).catch(() => {});
-             }
-
-             setText("");
-             toast({ title: "Reply posted successfully!" });
-
-        } catch (err: any) {
-            console.error("Error submitting reply:", err);
-            toast({ title: "Error", description: err.message || "Failed to post reply.", variant: "destructive" });
-        } finally {
-            setIsSubmitting(false);
-        }
-    };
-    
-    if (!user) {
-        return <p className="text-sm text-center text-muted-foreground py-4">Please sign in to reply.</p>
-    }
-
-    return (
-        <form id="reply-form" onSubmit={handleSubmit} className="flex items-start gap-2 pt-2">
-            <div className="flex-grow space-y-1">
-                <Textarea 
-                    name="text"
-                    value={text}
-                    onChange={(e) => setText(e.target.value)}
-                    placeholder="Add a reply... (Max 120 chars)" 
-                    maxLength={120} 
-                    rows={1} 
-                    required
-                    className="flex-grow min-h-0"
-                />
-                {error && <p className="text-sm text-destructive mt-1">{error}</p>}
-            </div>
-            <Button type="submit" size="icon" className="h-full" disabled={isSubmitting}>
-                 {isSubmitting ? <div className="animate-spin rounded-full h-4 w-4 border-b-2 border-current"></div> : <Send className="h-4 w-4" />}
-            </Button>
-        </form>
-    );
-}
-
-function ReportDialog({ note, open, onOpenChange, onReportSubmit }: { note: Note, open: boolean, onOpenChange: (open: boolean) => void, onReportSubmit: () => void }) {
-    const { user } = useAuth();
-    const { toast } = useToast();
-    const [reason, setReason] = useState('');
-    const [isSubmitting, setIsSubmitting] = useState(false);
-
-    const handleSubmit = async () => {
-        if (!user) {
-            toast({ title: "Not signed in", description: "You must be signed in to report.", variant: "destructive" });
-            return;
-        }
-        if (reason.length < 10) {
-            toast({ title: "Reason too short", description: "Please provide a more detailed reason.", variant: "destructive" });
-            return;
-        }
-
-        setIsSubmitting(true);
-        try {
-            await runTransaction(db, async (transaction) => {
-                const noteRef = doc(db, "notes", note.id);
-                const noteSnap = await transaction.get(noteRef);
-                const { newCount, hide } = calculateReportUpdate(noteSnap.data()?.reportCount ?? 0);
-                const updates: any = { reportCount: newCount };
-                if (hide) {
-                    updates.visibility = "unlisted";
-                }
-                transaction.update(noteRef, updates);
-
-                const reportRef = doc(collection(db, "reports"));
-                transaction.set(reportRef, {
-                    noteId: note.id,
-                    noteAuthorUid: note.authorUid,
-                    reporterUid: user.uid,
-                    reason: reason,
-                    createdAt: serverTimestamp(),
-                    status: "pending_review",
-                });
-            });
-
-            toast({
-                title: "Report Submitted",
-                description: "Thank you. The note has been flagged for review.",
-            });
-            onReportSubmit();
-        } catch (error: any) {
-            console.error("Error submitting report:", error);
-            toast({ title: "Error", description: error.message || "Failed to submit report.", variant: "destructive" });
-        } finally {
-            setIsSubmitting(false);
-        }
-    };
-
-    return (
-        <AlertDialog open={open} onOpenChange={onOpenChange}>
-            <AlertDialogContent>
-                <AlertDialogHeader>
-                    <AlertDialogTitle>Report Note</AlertDialogTitle>
-                    <AlertDialogDescription>
-                        Why are you reporting this note? Please provide a brief explanation. Your report is anonymous to other users.
-                    </AlertDialogDescription>
-                </AlertDialogHeader>
-                <div className="grid gap-2">
-                    <Label htmlFor="reason">Reason for reporting</Label>
-                    <Textarea
-                        id="reason"
-                        placeholder="e.g., This note contains harassment."
-                        value={reason}
-                        onChange={(e) => setReason(e.target.value)}
-                        maxLength={500}
-                        required
-                    />
-                </div>
-                <AlertDialogFooter>
-                    <AlertDialogCancel>Cancel</AlertDialogCancel>
-                    <AlertDialogAction onClick={handleSubmit} disabled={isSubmitting || reason.length < 10}>
-                        {isSubmitting ? 'Submitting...' : 'Submit Report'}
-                    </AlertDialogAction>
-                </AlertDialogFooter>
-            </AlertDialogContent>
-        </AlertDialog>
-    );
-}
-
-function NoteView({ note: initialNote, onClose }: {note: Note, onClose: () => void}) {
-    const { user } = useAuth();
-    const { toast } = useToast();
-    const [note, setNote] = useState<Note>(initialNote);
-    const [replies, setReplies] = useState<Reply[]>([]);
-    const [isLiked, setIsLiked] = useState(false);
-    const [isLiking, setIsLiking] = useState(false);
-    const [loadingReplies, setLoadingReplies] = useState(true);
-    const [isReportDialogOpen, setReportDialogOpen] = useState(false);
-    const [isDeleteDialogOpen, setDeleteDialogOpen] = useState(false);
-
-    useEffect(() => {
-        const noteRef = doc(db, 'notes', initialNote.id);
-        const unsubscribeNote = onSnapshot(noteRef, (doc) => {
-            if (doc.exists()) {
-                 const data = doc.data();
-                 if (data.visibility === 'unlisted' && data.authorUid !== user?.uid) {
-                    toast({title: "Note Unavailable", description: "This note is no longer available."});
-                    onClose();
-                    return;
-                 }
-                 const createdAt = data.createdAt instanceof Timestamp ? 
-                    { seconds: data.createdAt.seconds, nanoseconds: data.createdAt.nanoseconds } : 
-                    { seconds: Date.now() / 1000, nanoseconds: 0 };
-                setNote({ id: doc.id, ...data, createdAt } as Note);
-            } else {
-                toast({title: "Note Deleted", description: "This note has been removed."});
-                onClose();
-            }
-        }, (error) => {
-            console.error("Snapshot listener error:", error);
-            toast({title: "Network Error", description: "Could not sync with the note."});
-            onClose();
-        });
-        
-        return () => unsubscribeNote();
-    }, [initialNote.id, toast, onClose, user?.uid]);
-    
-    useEffect(() => {
-        if (!user) return;
-        const likeId = `${user.uid}_${initialNote.id}`;
-        const likeRef = doc(db, 'likes', likeId);
-        const unsubscribeLike = onSnapshot(likeRef, (doc) => {
-            setIsLiked(doc.exists());
-        });
-        return () => unsubscribeLike();
-    }, [initialNote.id, user]);
-
-    useEffect(() => {
-        const repliesRef = collection(db, 'notes', initialNote.id, 'replies');
-        const q = query(repliesRef, orderBy('createdAt', 'asc'));
-
-        const unsubscribe = onSnapshot(q, (querySnapshot) => {
-            const repliesData = querySnapshot.docs.map(doc => {
-                const data = doc.data();
-                const createdAt = data.createdAt instanceof Timestamp ? 
-                    { seconds: data.createdAt.seconds, nanoseconds: data.createdAt.nanoseconds } : 
-                    { seconds: Date.now() / 1000, nanoseconds: 0 };
-                return { id: doc.id, ...data, createdAt } as Reply;
-            });
-            setReplies(repliesData);
-            setLoadingReplies(false);
-        }, (error) => {
-            console.error("Error fetching replies:", error);
-            setLoadingReplies(false);
-        });
-
-        return () => unsubscribe();
-    }, [initialNote.id]);
-
-    const handleLikeToggle = async () => {
-        if (!user) {
-            toast({ title: "Please sign in to like notes.", variant: "destructive"});
-            return;
-        }
-        setIsLiking(true);
-
-      const noteRef = doc(db, 'notes', note.id);
-      const likeId = `${user.uid}_${note.id}`;
-      const likeRef = doc(db, 'likes', likeId);
-
-      let addedLike = false;
-      try {
-          await runTransaction(db, async (transaction) => {
-              const likeDoc = await transaction.get(likeRef);
-              const noteDoc = await transaction.get(noteRef);
-
-              if (!noteDoc.exists()) {
-                  throw new Error("Note does not exist!");
-              }
-
-              const currentScore = noteDoc.data().score || 0;
-
-              if (likeDoc.exists()) {
-                  transaction.delete(likeRef);
-                  transaction.update(noteRef, { score: currentScore - 1 });
-              } else {
-                  transaction.set(likeRef, {
-                      userId: user.uid,
-                      noteId: note.id,
-                      createdAt: serverTimestamp()
-                  });
-                  transaction.update(noteRef, { score: currentScore + 1 });
-                  addedLike = true;
-              }
-          });
-
-          if (addedLike && note.authorUid && note.authorUid !== user.uid) {
-              const pseudonym = await getOrCreatePseudonym(user.uid);
-              await addDoc(collection(db, 'notifications'), {
-                  userId: note.authorUid,
-                  type: 'like',
-                  noteId: note.id,
-                  actorUid: user.uid,
-                  actorPseudonym: pseudonym,
-                  createdAt: serverTimestamp(),
-                  read: false,
-              });
-              fetch('/api/notify', {
-                  method: 'POST',
-                  headers: { 'Content-Type': 'application/json' },
-                  body: JSON.stringify({
-                      userId: note.authorUid,
-                      title: 'New like',
-                      body: `${pseudonym} liked your note`,
-                      data: { noteId: note.id, type: 'like' }
-                  })
-              }).catch(() => {});
-          }
-      } catch (error: any) {
-          console.error("Transaction failed: ", error);
-          toast({ title: "Error", description: error.message || "Failed to update like status.", variant: "destructive" });
-      } finally {
-          setIsLiking(false);
-      }
-    };
-    
-    const handleDelete = async () => {
-        if (user?.uid !== note.authorUid) {
-            toast({ title: "Not Authorized", description: "You can only delete your own notes.", variant: "destructive" });
-            return;
-        }
-        try {
-            await deleteDoc(doc(db, "notes", note.id));
-            toast({ title: "Note Deleted", description: "Your note has been successfully deleted." });
-            onClose();
-        } catch (error: any) {
-            console.error("Error deleting note: ", error);
-            toast({ title: "Error", description: error.message || "Failed to delete the note.", variant: "destructive" });
-        }
-        setDeleteDialogOpen(false);
-    }
-
-    return (
-        <>
-            <SheetHeader>
-                <SheetTitle className="font-headline text-2xl">Note Revealed!</SheetTitle>
-            </SheetHeader>
-            <ScrollArea className="flex-1 -mx-6">
-                <div className="px-6 pt-4 space-y-6">
-                    {note.media?.[0]?.path && (
-                        <div className="rounded-lg overflow-hidden aspect-video relative bg-muted">
-                            <Image src={note.media[0].path} alt="Note media" fill className="object-cover" data-ai-hint="mural street art" />
-                        </div>
-                    )}
-                    <p className="text-lg leading-relaxed whitespace-pre-wrap">{note.text}</p>
-                    <div className="flex items-center justify-between text-muted-foreground text-sm">
-                        <div className="flex items-center gap-2">
-                            <Avatar className="w-8 h-8">
-                                <AvatarFallback>{note.authorPseudonym?.substring(0,2)}</AvatarFallback>
-                            </Avatar>
-                            <span>{note.authorPseudonym}</span>
-                        </div>
-                        <span>{note.createdAt ? new Date(note.createdAt.seconds * 1000).toLocaleDateString() : 'Just now'}</span>
-                    </div>
-                    <div className="flex items-center gap-2 flex-wrap">
-                        <Button variant="outline" size="sm" onClick={handleLikeToggle} disabled={isLiking || note.authorUid === user?.uid}>
-                            <Heart className={cn("h-4 w-4 mr-2", isLiked && "fill-destructive text-destructive")} /> 
-                            {note.score}
-                        </Button>
-                        <Button variant="outline" size="sm" onClick={() => setReportDialogOpen(true)} disabled={note.authorUid === user?.uid}><Flag className="h-4 w-4 mr-2"/> Report</Button>
-                        {note.authorUid === user?.uid && (
-                            <Button variant="destructive" size="sm" onClick={() => setDeleteDialogOpen(true)}>
-                                <Trash2 className="h-4 w-4 mr-2" />
-                                Delete
-                            </Button>
-                        )}
-                        <Badge variant={note.type === 'photo' ? 'default' : 'secondary'}>{note.type}</Badge>
-                    </div>
-                    <Separator />
-                    <div className="space-y-4">
-                        <h3 className="font-headline text-xl">Replies ({replies.length})</h3>
-                        <div className="space-y-4">
-                            {loadingReplies && <Skeleton className="h-16 w-full" />}
-                            {replies.map(reply => (
-                                <div key={reply.id} className="flex gap-3">
-                                    <Avatar className="h-8 h-8">
-                                        <AvatarFallback>{reply.authorPseudonym?.substring(0,2)}</AvatarFallback>
-                                    </Avatar>
-                                    <div className="bg-muted p-3 rounded-lg flex-1">
-                                        <p className="text-sm whitespace-pre-wrap">{reply.text}</p>
-                                        <p className="text-xs text-muted-foreground mt-1">{reply.authorPseudonym} &middot; {new Date(reply.createdAt.seconds * 1000).toLocaleDateString()}</p>
-                                    </div>
-                                </div>
-                            ))}
-                            {!loadingReplies && replies.length === 0 && <p className="text-sm text-muted-foreground text-center">Be the first to reply!</p>}
-                        </div>
-                    </div>
-                </div>
-            </ScrollArea>
-            <SheetFooter className="pt-2">
-                <ReplyForm noteId={note.id} noteAuthorUid={note.authorUid} />
-            </SheetFooter>
-
-            <ReportDialog 
-                note={note} 
-                open={isReportDialogOpen} 
-                onOpenChange={setReportDialogOpen}
-                onReportSubmit={() => {
-                    setReportDialogOpen(false);
-                    onClose();
-                }}
-            />
-             <AlertDialog open={isDeleteDialogOpen} onOpenChange={setDeleteDialogOpen}>
-                <AlertDialogContent>
-                    <AlertDialogHeader>
-                        <AlertDialogTitle>Are you absolutely sure?</AlertDialogTitle>
-                        <AlertDialogDescription>
-                            This action cannot be undone. This will permanently delete your note and all of its replies.
-                        </AlertDialogDescription>
-                    </AlertDialogHeader>
-                    <AlertDialogFooter>
-                        <AlertDialogCancel>Cancel</AlertDialogCancel>
-                        <AlertDialogAction onClick={handleDelete}>Continue</AlertDialogAction>
-                    </AlertDialogFooter>
-                </AlertDialogContent>
-            </AlertDialog>
-        </>
-    )
-}
+import { Skeleton } from "./ui/skeleton";
+import CreateNoteForm from "./create-note-form";
+import NoteView from "./note-view";
 
 interface NoteSheetContentProps {
   noteId: string | null;
@@ -688,7 +20,13 @@ interface NoteSheetContentProps {
   onClose: () => void;
 }
 
-export default function NoteSheetContent({ noteId, isCreating, userLocation, onNoteCreated, onClose }: NoteSheetContentProps) {
+export default function NoteSheetContent({
+  noteId,
+  isCreating,
+  userLocation,
+  onNoteCreated,
+  onClose,
+}: NoteSheetContentProps) {
   const { toast } = useToast();
   const [note, setNote] = useState<Note | null>(null);
   const [loadingNote, setLoadingNote] = useState(false);
@@ -698,12 +36,13 @@ export default function NoteSheetContent({ noteId, isCreating, userLocation, onN
       if (noteId && !isCreating) {
         setLoadingNote(true);
         try {
-          const noteDoc = await getDoc(doc(db, 'notes', noteId));
+          const noteDoc = await getDoc(doc(db, "notes", noteId));
           if (noteDoc.exists()) {
             const data = noteDoc.data();
-            const createdAt = data.createdAt instanceof Timestamp ? 
-                { seconds: data.createdAt.seconds, nanoseconds: data.createdAt.nanoseconds } : 
-                { seconds: Date.now() / 1000, nanoseconds: 0 };
+            const createdAt =
+              data.createdAt instanceof Timestamp
+                ? { seconds: data.createdAt.seconds, nanoseconds: data.createdAt.nanoseconds }
+                : { seconds: Date.now() / 1000, nanoseconds: 0 };
             setNote({ id: noteDoc.id, ...data, createdAt } as Note);
           } else {
             toast({ title: "Error", description: "Could not find the selected note.", variant: "destructive" });
@@ -722,7 +61,6 @@ export default function NoteSheetContent({ noteId, isCreating, userLocation, onN
     }
     fetchNote();
   }, [noteId, isCreating, toast, onClose]);
-  
 
   if (isCreating) {
     return <CreateNoteForm userLocation={userLocation} onNoteCreated={onNoteCreated} onClose={onClose} />;
@@ -730,15 +68,15 @@ export default function NoteSheetContent({ noteId, isCreating, userLocation, onN
 
   if (loadingNote) {
     return (
-        <div className="p-4 space-y-6">
-            <Skeleton className="h-64 w-full rounded-lg" />
-            <Skeleton className="h-6 w-3/4" />
-            <Skeleton className="h-4 w-1/2" />
-            <div className="flex items-center justify-between">
-                <Skeleton className="h-10 w-24" />
-                <Skeleton className="h-10 w-24" />
-            </div>
+      <div className="p-4 space-y-6">
+        <Skeleton className="h-64 w-full rounded-lg" />
+        <Skeleton className="h-6 w-3/4" />
+        <Skeleton className="h-4 w-1/2" />
+        <div className="flex items-center justify-between">
+          <Skeleton className="h-10 w-24" />
+          <Skeleton className="h-10 w-24" />
         </div>
+      </div>
     );
   }
 
@@ -746,6 +84,4 @@ export default function NoteSheetContent({ noteId, isCreating, userLocation, onN
 
   return <NoteView note={note} onClose={onClose} />;
 }
-    
 
-
